requirements.txt
google-cloud-asset==3.26.1
python-dotenv